+++ src/lib/components/FrontPage.svelte
@@ -1,4 +1,6 @@
 <script lang="ts">
+	import { checkHexa } from '$lib/ts/checkHexa';
+
 	let inputfield: string = '';
 
 	$: {
@@ -19,19 +21,6 @@
 
 			window.location.replace(url);
 		}
-	}
-
-	function checkHexa(input: string) {
-		for (let i = 0; i < input.length; i++) {
-			let charCode = input.charCodeAt(i);
-
-			//check if char is between 0 and 9 or A and F
-			if (!(charCode >= 48 && charCode <= 57) && !(charCode >= 65 && charCode <= 70)) {
-				return false;
-			}
-		}
-
-		return true;
 	}
 </script>
