+++ src/routes/tests/unitTests.ts
@@ -1,0 +1,7 @@
+import {describe, test, expect } from 'vitest';
+
+describe("Example File", () => {
+    test("Sample test", () => {
+        expect(1 + 3).equal(4);
+    });
+});
