+++ vite.config.ts
@@ -1,9 +1,11 @@
-import { sveltekit } from '@sveltejs/kit/vite';
-import { defineConfig } from 'vitest/config';
+import { defineConfig } from 'vitest/config'
+import { svelte } from '@sveltejs/vite-plugin-svelte'
 
 export default defineConfig({
-	plugins: [sveltekit()],
-	test: {
-		include: ['src/**/*.{test,spec}.{js,ts}']
-	}
-});
+  plugins: [
+    svelte({ hot: !process.env.VITEST }),
+  ],
+  test: {
+    environment: 'jsdom',
+  }
+})
